chore: rename core folder to vee-validate
