+++ packages/i18n/tests/index.spec.ts
@@ -1,8 +1,8 @@
 import flushPromises from 'flush-promises';
-import { defineRule, configure } from '@/core';
+import { defineRule, configure } from '@/vee-validate';
 import { required, between } from '@/rules';
 import { localize, setLocale } from '@/i18n';
-import { mountWithHoc, setValue } from '../../core/tests/helpers';
+import { mountWithHoc, setValue } from '../../vee-validate/tests/helpers';
 
 defineRule('required', required);
 defineRule('between', between);
