+++ packages/vee-validate/src/symbols.ts
