+++ packages/vee-validate/tests/ErrorMessage.spec.ts
@@ -1,5 +1,5 @@
 import flushPromises from 'flush-promises';
-import { defineRule, ErrorMessage } from '@/core';
+import { defineRule, ErrorMessage } from '@/vee-validate';
 import { mountWithHoc, setValue } from './helpers';
 
 describe('<ErrorMessage />', () => {
