+++ packages/vee-validate/tests/Form.spec.ts
@@ -1,5 +1,5 @@
 import flushPromises from 'flush-promises';
-import { defineRule } from '@/core';
+import { defineRule } from '@/vee-validate';
 import { mountWithHoc, setValue, setChecked } from './helpers';
 import * as yup from 'yup';
 import { onErrorCaptured, reactive, ref, Ref } from 'vue';
